+++ handler/dbsvr.py
@@ -10,7 +10,7 @@ import mlib_iisi.utils as libiisi
 import mxweb
 import mxpsu as mx
 from tornado import gen
-from tornado.httpclient import AsyncHTTPClient
+# from tornado.httpclient import AsyncHTTPClient
 import base
 import time
 import json
@@ -24,7 +24,8 @@ class DBSvrHandler(base.RequestHandler):
     &nbsp;&nbsp;参考中间层接口相关文档'''
 
     keep_name_case = False
-    thc = AsyncHTTPClient()
+    # thc = AsyncHTTPClient()
+
     root_path = r'/dbsvr/'
 
     @gen.coroutine
@@ -33,10 +34,10 @@ class DBSvrHandler(base.RequestHandler):
                                self.request.uri.replace(self.root_path, ''))
         try:
             # rep = utils.m_httpclinet_pool.request('GET', url, fields={}, timeout=7.0, retries=False)
-
-            rep = yield self.thc.fetch(
-                url, raise_error=True, request_timeout=12)
-            self.write(rep.body)
+            rep = self._pm.request('GET', url, timeout=10.0)
+            # rep = yield self.thc.fetch(
+            #     url, raise_error=True, request_timeout=12)
+            self.write(rep.data)
         except Exception as ex:
             self.write(str(ex))
         self.finish()
@@ -63,13 +64,14 @@ class DBSvrHandler(base.RequestHandler):
         # except Exception as ex:
         #     self.write(str(ex))
         try:
-            rep = yield self.thc.fetch(
-                url,
-                method="POST",
-                body=json.dumps(data),
-                raise_error=True,
-                request_timeout=10)
-            self.write(rep.body)
+            rep = self._pm.request('POST', url, fields=data, timeout=10.0)
+            # rep = yield self.thc.fetch(
+            #     url,
+            #     method="POST",
+            #     body=json.dumps(data),
+            #     raise_error=True,
+            #     request_timeout=10)
+            self.write(rep.data)
         except Exception as ex:
             self.write(str(ex))
         self.finish()